# gen/openapiv2/booking/v1/booking.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "booking/v1/booking.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BookingService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/booking/orders/{orderId}:cancel": {
      "post": {
        "operationId": "BookingService_CancelOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CancelOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                },
                "idempotencyKey": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    },
    "/v1/booking/orders:checkout": {
      "post": {
        "operationId": "BookingService_CheckoutOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CheckoutOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CheckoutOrderRequest"
            }
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    },
    "/v1/booking/reservations": {
      "post": {
        "operationId": "BookingService_ReserveSeats",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ReserveSeatsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ReserveSeatsRequest"
            }
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CancelOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "releasedSeatCount": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1CheckoutOrderRequest": {
      "type": "object",
      "properties": {
        "reservationId": {
          "type": "string"
        },
        "paymentMethodToken": {
          "type": "string"
        },
        "idempotencyKey": {
          "type": "string"
        }
      }
    },
    "v1CheckoutOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "paymentTransactionId": {
          "type": "string"
        }
      }
    },
    "v1ReserveSeatsRequest": {
      "type": "object",
      "properties": {
        "eventId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "seatIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "idempotencyKey": {
          "type": "string"
        }
      }
    },
    "v1ReserveSeatsResponse": {
      "type": "object",
      "properties": {
        "reservationId": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time"
        },
        "totalAmountMinor": {
          "type": "string",
          "format": "int64"
        },
        "currency": {
          "type": "string"
        }
      }
    }
  }
}